added new pool threshold from blockchain

See merge request elixxir/registration!23

+++ scheduling/params.go
@@ -58,9 +58,8 @@ type Params struct {
 	DebugTrackRounds bool
 
 	//SECURE ONLY
-	// sets the minimum number of nodes in the waiting pool before secure teaming
-	// wil create a team
-	Threshold uint32
+	// Minimum percentage of nodes in the waiting pool before secure teaming wil create a team
+	Threshold float64
 }
 
 //internal structure which describes a round to be created
